+++ base/VkSemaphoreBundle.cpp
@@ -19,12 +19,12 @@ VkSemaphoreBundle::VkSemaphoreBundle(const VkDeviceManager &_device_manager, uin
 	}
 }
 
-VkSemaphore VkSemaphoreBundle::operator[](uint32_t num) const
+VkSemaphore VkSemaphoreBundle::operator[](size_t num) const
 {
 	return syn_objects_bundle[num];
 }
 
-const VkSemaphore &VkSemaphoreBundle::GetOne(uint32_t num) const
+const VkSemaphore &VkSemaphoreBundle::GetOne(size_t num) const
 {
 	return syn_objects_bundle[num];
 }
